refeactor: rename PackageManager
