+++ lib/PackageManager.js
